--- a/src/torematrix/ui/utils/__init__.py
+++ b/src/torematrix/ui/utils/__init__.py
@@ -1,0 +1,6 @@
+"""UI utilities package for ToreMatrix V3."""
+
+from .dpi_utils import DPIManager
+from .layout_utils import LayoutUtilities
+
+__all__ = ['DPIManager', 'LayoutUtilities']