--- a/tests/unit/ui/layouts/__init__.py
+++ b/tests/unit/ui/layouts/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for ToreMatrix Layout Management System."""