--- a/tests/unit/ui/themes/__init__.py
+++ b/tests/unit/ui/themes/__init__.py
@@ -1,0 +1,1 @@
+"""Unit tests for the theme framework."""