… CLAUDE.md
… THEME_FRAMEWORK_API.md
… src/torematrix/core/models/__pycache__/__init__.cpython-312.pyc
… src/torematrix/core/models/__pycache__/base_types.cpython-312.pyc
… src/torematrix/core/models/__pycache__/coordinates.cpython-312.pyc
… src/torematrix/core/models/__pycache__/element.cpython-312.pyc
… src/torematrix/core/models/__pycache__/metadata.cpython-312.pyc
… src/torematrix/ui/benchmarks.py
… src/torematrix/ui/layouts/__init__.py
… src/torematrix/ui/layouts/base.py
… src/torematrix/ui/layouts/manager.py
… src/torematrix/ui/layouts/templates.py
… src/torematrix/ui/layouts/validation.py
… src/torematrix/ui/performance.py
… src/torematrix/ui/styles/dark.qss
… src/torematrix/ui/styles/light.qss
… src/torematrix/ui/themes/README.md
… src/torematrix/ui/themes/base.py
… src/torematrix/ui/themes/engine.py
… src/torematrix/ui/themes/exceptions.py
… src/torematrix/ui/themes/types.py
… src/torematrix/ui/utils/dpi_utils.py
… src/torematrix/ui/utils/layout_utils.py
… test_reactive_components.py
… tests/unit/ui/components/test_decorators.py
… tests/unit/ui/components/test_decorators_isolated.py
… tests/unit/ui/components/test_lifecycle.py
… tests/unit/ui/components/test_lifecycle_isolated.py
… tests/unit/ui/components/test_reactive.py
… tests/unit/ui/components/test_reactive_isolated.py